feat: Add an option to include minimized windows when restoring all #46
