+++ packaging/WindowResizer.Packaging/Package.appxmanifest
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 
 <Package
   xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
@@ -10,7 +10,7 @@
     <Identity
       Name="9577ColdMoon.WindowResizer"
       Publisher="CN=007FCD3A-16C6-4DE3-B752-6CFE9BC6BF39"
-      Version="1.3.2.0" />
+      Version="1.3.3.0" />
 
     <Properties>
         <DisplayName>WindowResizer</DisplayName>
